requirements.txt
# 核心依赖
requests>=2.28.0
pyyaml>=6.0
ccxt>=4.0.0
python-dateutil>=2.8.0

# 数据处理
decimal
typing-extensions

# 安全和加密
cryptography>=3.4.0

# 可选依赖（用于测试和开发）
pytest>=7.0.0
flake8>=5.0.0
black>=22.0.0